+++ src/main/java/com/chirag/exceltomysql/repository/ProductRepository.java
@@ -4,5 +4,5 @@ package com.chirag.exceltomysql.repository;
 import com.chirag.exceltomysql.entity.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Products, Long> {
+public interface ProductRepository extends JpaRepository<Products, Double> {
 }
